refactor(binder): isolate raw cx bindings inside server
