+++ shared/src/types.ts
@@ -76,28 +76,6 @@ export type StyleImport =
     };
 
 /**
- * A single `const cx = classNames.bind(styles)` binding detected
- * in one source file. A file may have several bindings (different
- * `styles` imports, different scopes).
- */
-export interface CxBinding {
-  /** Identifier used at call sites — `cx`, `classes`, `cxBtn`, etc. */
-  readonly cxVarName: string;
-  /** Identifier for the styles default-import. */
-  readonly stylesVarName: string;
-  /** Absolute path of the `.module.scss|css` file the binding resolves to. */
-  readonly scssModulePath: string;
-  /** Source range covering the binding identifier declaration. */
-  readonly bindingRange: Range;
-  /**
-   * Identifier the `classnames/bind` default import was bound to
-   * in this file. Usually `"classNames"`, but aliased imports allow any
-   * name (e.g. `"cn"`).
-   */
-  readonly classNamesImportName: string;
-}
-
-/**
  * Which source syntax produced a class reference.
  * Deliberately a string literal union (no payload); this can be
  * widened to a discriminated union later without breaking
